+++ src/main/java/fll/scheduler/ScheduleChecker.java
@@ -38,7 +38,7 @@ public class ScheduleChecker {
 
   /**
    * Initialize a scheduler checker.
-   * 
+   *
    * @param params the parameters to use when checking the schedule
    * @param schedule the schedule to check
    */
@@ -53,7 +53,7 @@ public class ScheduleChecker {
       for (final Map.Entry<String, List<TeamScheduleInfo>> timeEntry : dateEntry.getValue().entrySet()) {
         final List<TeamScheduleInfo> tableMatches = timeEntry.getValue();
         if (tableMatches.size() > 2) {
-          final List<Integer> teams = new LinkedList<Integer>();
+          final List<Integer> teams = new LinkedList<>();
           for (final TeamScheduleInfo team : tableMatches) {
             teams.add(team.getTeamNumber());
           }
@@ -75,13 +75,13 @@ public class ScheduleChecker {
     // constraint: tournament:1
     final Map<LocalTime, Set<TeamScheduleInfo>> teamsAtTime = new HashMap<>();
     for (final TeamScheduleInfo si : schedule.getSchedule()) {
-      for (int round = 0; round < schedule.getNumberOfRounds(); ++round) {
+      for (final PerformanceTime performance : si.getAllPerformances()) {
         final Set<TeamScheduleInfo> teams;
-        if (teamsAtTime.containsKey(si.getPerfTime(round))) {
-          teams = teamsAtTime.get(si.getPerfTime(round));
+        if (teamsAtTime.containsKey(performance.getTime())) {
+          teams = teamsAtTime.get(performance.getTime());
         } else {
-          teams = new HashSet<TeamScheduleInfo>();
-          teamsAtTime.put(si.getPerfTime(round), teams);
+          teams = new HashSet<>();
+          teamsAtTime.put(performance.getTime(), teams);
         }
         teams.add(si);
       }
@@ -118,7 +118,7 @@ public class ScheduleChecker {
         if (teamsAtTime.containsKey(subj.getTime())) {
           teams = teamsAtTime.get(subj.getTime());
         } else {
-          teams = new HashSet<TeamScheduleInfo>();
+          teams = new HashSet<>();
           teamsAtTime.put(subj.getTime(), teams);
         }
         teams.add(si);
@@ -135,7 +135,7 @@ public class ScheduleChecker {
                                                  new SubjectiveTime(topEntry.getKey(), entry.getKey()), null, message));
         }
 
-        final Set<String> judges = new HashSet<String>();
+        final Set<String> judges = new HashSet<>();
         for (final TeamScheduleInfo ti : entry.getValue()) {
           if (!judges.add(ti.getJudgingGroup())) {
             final String message = String.format("%s judge %s cannot see more than one team at %s", topEntry.getKey(),
@@ -199,34 +199,33 @@ public class ScheduleChecker {
     checkConstraintTeam3(violations, ti);
 
     // constraint: team:4
-    for (int round = 0; round < schedule.getNumberOfRounds(); ++round) {
-      final String performanceName = String.valueOf(round
-          + 1);
-      verifyPerformanceVsSubjective(violations, ti, performanceName, ti.getPerfTime(round));
+    for (final PerformanceTime performance : ti.getAllPerformances()) {
+      final String performanceName = ti.getRoundName(performance);
+      verifyPerformanceVsSubjective(violations, ti, performanceName, performance.getTime());
     }
 
     // constraint: team:5
-    final Map<String, PerformanceTime> tableSides = new HashMap<String, PerformanceTime>();
-    for (int round = 0; round < schedule.getNumberOfRounds(); ++round) {
-      final TeamScheduleInfo opponent = schedule.findOpponent(ti, round);
+    final Map<String, PerformanceTime> tableSides = new HashMap<>();
+    for (final PerformanceTime performance : ti.getAllPerformances()) {
+      final TeamScheduleInfo opponent = schedule.findOpponent(ti, performance);
       if (null != opponent) {
         if (opponent.getTeamNumber() == ti.getTeamNumber()) {
           throw new FLLRuntimeException("Internal error, findOpponent is broken and returned the same team");
         }
 
-        final String key = ti.getPerfTableColor(round)
+        final String key = performance.getTable()
             + " "
-            + ti.getPerfTableSide(round);
+            + performance.getSide();
         if (tableSides.containsKey(key)) {
-          final PerformanceTime otherTime = tableSides.get(key);
+          final PerformanceTime otherPerformance = tableSides.get(key);
           final String tableMessage = String.format("Team %d is competing on %s %d more than once", ti.getTeamNumber(),
-                                                    ti.getPerfTableColor(round), ti.getPerfTableSide(round));
+                                                    performance.getTable(), performance.getSide());
           violations.add(new ConstraintViolation(ConstraintViolation.Type.SOFT, ti.getTeamNumber(), null, null,
-                                                 ti.getPerfTime(round), tableMessage));
+                                                 performance.getTime(), tableMessage));
           violations.add(new ConstraintViolation(ConstraintViolation.Type.SOFT, ti.getTeamNumber(), null, null,
-                                                 otherTime.getTime(), tableMessage));
+                                                 otherPerformance.getTime(), tableMessage));
         } else {
-          tableSides.put(key, ti.getPerf(round));
+          tableSides.put(key, performance);
         }
 
         /*
@@ -242,54 +241,50 @@ public class ScheduleChecker {
 
         int opponentSide = -1;
         // figure out which round matches up
-        for (int oround = 0; oround < schedule.getNumberOfRounds(); ++oround) {
-          if (opponent.getPerfTime(oround).equals(ti.getPerfTime(round))) {
-            opponentSide = opponent.getPerfTableSide(oround);
+        for (final PerformanceTime opponentPerformance : opponent.getAllPerformances()) {
+          if (opponentPerformance.getTime().equals(performance.getTime())) {
+            opponentSide = opponentPerformance.getSide();
             break;
           }
         }
         if (-1 == opponentSide) {
           final String message = String.format("Unable to find time match for rounds between team %d and team %d at time %s",
                                                ti.getTeamNumber(), opponent.getTeamNumber(),
-                                               TournamentSchedule.formatTime(ti.getPerfTime(round)));
+                                               TournamentSchedule.formatTime(performance.getTime()));
           violations.add(new ConstraintViolation(ConstraintViolation.Type.HARD, ti.getTeamNumber(), null, null,
-                                                 ti.getPerfTime(round), message));
+                                                 performance.getTime(), message));
         } else {
-          if (opponentSide == ti.getPerfTableSide(round)) {
-            final String message = String.format("Team %d and team %d are both on table %s side %d at the same time for round %d",
-                                                 ti.getTeamNumber(), opponent.getTeamNumber(),
-                                                 ti.getPerfTableColor(round), ti.getPerfTableSide(round), (round
-                                                     + 1));
+          if (opponentSide == performance.getSide()) {
+            final String message = String.format("Team %d and team %d are both on table %s side %d at the same time for round %s",
+                                                 ti.getTeamNumber(), opponent.getTeamNumber(), performance.getTable(),
+                                                 performance.getSide(), ti.getRoundName(performance));
             violations.add(new ConstraintViolation(ConstraintViolation.Type.HARD, ti.getTeamNumber(), null, null,
-                                                   ti.getPerfTime(round), message));
+                                                   performance.getTime(), message));
           }
         }
 
-        for (int r = round
-            + 1; r < schedule.getNumberOfRounds(); ++r) {
-          final TeamScheduleInfo otherOpponent = schedule.findOpponent(ti, r);
+        ti.getAllPerformances().stream().filter(p -> p.getTime().isAfter(performance.getTime())).forEach(p -> {
+          final TeamScheduleInfo otherOpponent = schedule.findOpponent(ti, p);
           if (otherOpponent != null
               && opponent.equals(otherOpponent)) {
-            final String message = String.format("Team %d competes against %d more than once rounds: %d, %d",
-                                                 ti.getTeamNumber(), opponent.getTeamNumber(), (round
-                                                     + 1),
-                                                 (r
-                                                     + 1));
+            final String message = String.format("Team %d competes against %d more than once rounds: %s, %s",
+                                                 ti.getTeamNumber(), opponent.getTeamNumber(),
+                                                 ti.getRoundName(performance), ti.getRoundName(p));
             violations.add(new ConstraintViolation(ConstraintViolation.Type.SOFT, ti.getTeamNumber(), null, null,
-                                                   ti.getPerfTime(round), message));
+                                                   performance.getTime(), message));
             violations.add(new ConstraintViolation(ConstraintViolation.Type.SOFT, ti.getTeamNumber(), null, null,
-                                                   ti.getPerfTime(r), message));
+                                                   p.getTime(), message));
             violations.add(new ConstraintViolation(ConstraintViolation.Type.SOFT, opponent.getTeamNumber(), null, null,
-                                                   ti.getPerfTime(round), message));
+                                                   performance.getTime(), message));
             violations.add(new ConstraintViolation(ConstraintViolation.Type.SOFT, opponent.getTeamNumber(), null, null,
-                                                   ti.getPerfTime(r), message));
+                                                   p.getTime(), message));
           }
-        }
+        });
       } else {
-        final String message = String.format("Team %d has no opponent for round %d", ti.getTeamNumber(), (round
-            + 1));
+        final String message = String.format("Team %d has no opponent for round %d", ti.getTeamNumber(),
+                                             ti.getRoundName(performance));
         violations.add(new ConstraintViolation(ConstraintViolation.Type.SOFT, ti.getTeamNumber(), null, null,
-                                               ti.getPerfTime(round), message));
+                                               performance.getTime(), message));
       }
     }
   }
@@ -299,48 +294,39 @@ public class ScheduleChecker {
    */
   private void checkConstraintTeam3(final Collection<ConstraintViolation> violations,
                                     final TeamScheduleInfo ti) {
-    if (schedule.getNumberOfRounds() < 2) {
+    if (schedule.getTotalNumberOfRounds() < 2) {
       // nothing to check
       return;
     }
 
-    for (int round = 1; round < schedule.getNumberOfRounds(); ++round) {
-      final int prevRound = round
-          - 1;
-      final Duration perfChangetime;
-      final int prevRoundOpponentRound = schedule.findOpponentRound(ti, prevRound);
-      final int curRoundOpponentRound = schedule.findOpponentRound(ti, round);
-      if (prevRoundOpponentRound != prevRound
-          || curRoundOpponentRound != round) {
-        perfChangetime = getSpecialPerformanceChangetime();
-      } else {
-        perfChangetime = getPerformanceChangetime();
-      }
+    PerformanceTime prevPerformance = null;
+    for (final PerformanceTime performance : ti.getAllPerformances()) {
+      if (null != prevPerformance) {
 
-      final LocalTime prevRoundStart = ti.getPerfTime(prevRound);
-      final LocalTime prevRoundEnd = prevRoundStart.plus(getPerformanceDuration());
-      final LocalTime roundStart = ti.getPerfTime(round);
-      if (prevRoundEnd.isAfter(roundStart)) {
-        final String message = String.format("Team %d is still in performance %d when they are to start performance %d: %s - %s",
-                                             ti.getTeamNumber(), prevRound
-                                                 + 1,
-                                             round
-                                                 + 1,
-                                             TournamentSchedule.formatTime(ti.getPerfTime(prevRound)),
-                                             TournamentSchedule.formatTime(ti.getPerfTime(round)));
-        violations.add(new ConstraintViolation(ConstraintViolation.Type.HARD, ti.getTeamNumber(), null, null,
-                                               ti.getPerfTime(round), message));
-      } else if (prevRoundEnd.plus(perfChangetime).isAfter(roundStart)) {
-        final String message = String.format("Team %d doesn't have enough time (%s) between performance %d and performance %d: %s - %s",
-                                             ti.getTeamNumber(), perfChangetime.toString(), prevRound
-                                                 + 1,
-                                             round
-                                                 + 1,
-                                             TournamentSchedule.formatTime(ti.getPerfTime(prevRound)),
-                                             TournamentSchedule.formatTime(ti.getPerfTime(round)));
-        violations.add(new ConstraintViolation(ConstraintViolation.Type.HARD, ti.getTeamNumber(), null, null,
-                                               ti.getPerfTime(round), message));
+        final Duration perfChangetime = getPerformanceChangetime();
+
+        final LocalTime prevRoundStart = prevPerformance.getTime();
+        final LocalTime prevRoundEnd = prevRoundStart.plus(getPerformanceDuration());
+        final LocalTime roundStart = performance.getTime();
+        if (prevRoundEnd.isAfter(roundStart)) {
+          final String message = String.format("Team %d is still in performance %s when they are to start performance %s: %s - %s", //
+                                               ti.getTeamNumber(), ti.getRoundName(prevPerformance),
+                                               ti.getRoundName(performance),
+                                               TournamentSchedule.formatTime(prevPerformance.getTime()),
+                                               TournamentSchedule.formatTime(performance.getTime()));
+          violations.add(new ConstraintViolation(ConstraintViolation.Type.HARD, ti.getTeamNumber(), null, null,
+                                                 performance.getTime(), message));
+        } else if (prevRoundEnd.plus(perfChangetime).isAfter(roundStart)) {
+          final String message = String.format("Team %d doesn't have enough time (%s) between performance %s and performance %s: %s - %s", //
+                                               ti.getTeamNumber(), ti.getRoundName(prevPerformance),
+                                               ti.getRoundName(performance), perfChangetime.toString(),
+                                               TournamentSchedule.formatTime(prevPerformance.getTime()),
+                                               TournamentSchedule.formatTime(performance.getTime()));
+          violations.add(new ConstraintViolation(ConstraintViolation.Type.HARD, ti.getTeamNumber(), null, null,
+                                                 performance.getTime(), message));
+        }
       }
+      prevPerformance = performance;
     }
   }
 
@@ -385,12 +371,12 @@ public class ScheduleChecker {
 
   /**
    * Verify the schedule.
-   * 
+   *
    * @return the constraint violations found, empty if no violations
    * @throws IOException
    */
   public List<ConstraintViolation> verifySchedule() {
-    final List<ConstraintViolation> constraintViolations = new LinkedList<ConstraintViolation>();
+    final List<ConstraintViolation> constraintViolations = new LinkedList<>();
 
     for (final TeamScheduleInfo verify : schedule.getSchedule()) {
       verifyTeam(constraintViolations, verify);
@@ -406,7 +392,7 @@ public class ScheduleChecker {
 
   /**
    * Make sure that there are no overlaps in times at each scheduling location.
-   * 
+   *
    * @param constraintViolations
    */
   private void verifyNoOverlap(final List<ConstraintViolation> violations) {
@@ -423,15 +409,16 @@ public class ScheduleChecker {
         }
         SortedSet<LocalTime> times = judges.get(judge);
         if (null == times) {
-          times = new TreeSet<LocalTime>();
+          times = new TreeSet<>();
           judges.put(judge, times);
         }
         times.add(subj.getTime());
       }
 
-      for (int round = 0; round < schedule.getNumberOfRounds(); ++round) {
-        final String table = si.getPerfTableColor(round);
-        final int side = si.getPerfTableSide(round);
+      for (final PerformanceTime performanceTime : si.getAllPerformances()) {
+
+        final String table = performanceTime.getTable();
+        final int side = performanceTime.getSide();
         final String tableKey = table
             + " "
             + side;
@@ -440,7 +427,7 @@ public class ScheduleChecker {
           performance = new TreeSet<>();
           tableToTime.put(tableKey, performance);
         }
-        performance.add(si.getPerfTime(round));
+        performance.add(performanceTime.getTime());
       }
     }
 
@@ -489,7 +476,7 @@ public class ScheduleChecker {
 
   /**
    * Time to allocate for a performance run.
-   * 
+   *
    * @return the performanceDuration
    */
   public Duration getPerformanceDuration() {
@@ -498,7 +485,7 @@ public class ScheduleChecker {
 
   /**
    * Time to allocate for either subjective judging.
-   * 
+   *
    * @return the subjectiveDuration (milliseconds)
    */
   public Duration getSubjectiveDuration(final String name) {
@@ -508,7 +495,7 @@ public class ScheduleChecker {
 
   /**
    * This is the time required between events.
-   * 
+   *
    * @return the changetime
    */
   public Duration getChangetime() {
@@ -526,7 +513,7 @@ public class ScheduleChecker {
    * This is the time required between performance runs for the two teams in
    * involved in the performance run that crosses round 1 and round 2 when there
    * is an odd number of teams.
-   * 
+   *
    * @return the specialPerformanceChangetime
    */
   public Duration getSpecialPerformanceChangetime() {