+++ src/main/java/fll/scheduler/SchedulerTableModel.java
@@ -6,13 +6,15 @@
 
 package fll.scheduler;
 
+import java.time.LocalTime;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
-import java.util.Date;
 import java.util.List;
 
 import javax.swing.table.AbstractTableModel;
+
+import org.apache.commons.lang3.tuple.Pair;
 
 /**
  * Table model for scheduler information used in {@link SchedulerUI}.
@@ -25,33 +27,33 @@ import javax.swing.table.AbstractTableModel;
 
   private final List<String> subjectiveColumns;
 
-  private static final Comparator<TeamScheduleInfo> TEAM_NUMBER_COMPARATOR = new Comparator<TeamScheduleInfo>() {
-    public int compare(final TeamScheduleInfo one,
-                       final TeamScheduleInfo two) {
-      if (one.getTeamNumber() < two.getTeamNumber()) {
-        return -1;
-      } else if (one.getTeamNumber() > two.getTeamNumber()) {
-        return 1;
-      } else {
-        return 0;
-      }
+  private static final Comparator<TeamScheduleInfo> TEAM_NUMBER_COMPARATOR = (one,
+                                                                              two) -> {
+    if (one.getTeamNumber() < two.getTeamNumber()) {
+      return -1;
+    } else if (one.getTeamNumber() > two.getTeamNumber()) {
+      return 1;
+    } else {
+      return 0;
     }
   };
 
   public SchedulerTableModel(final TournamentSchedule schedule) {
     this.schedule = schedule;
-    this.scheduleData = new ArrayList<TeamScheduleInfo>(schedule.getSchedule());
-    this.subjectiveColumns = new ArrayList<String>(schedule.getSubjectiveStations());
+    this.scheduleData = new ArrayList<>(schedule.getSchedule());
+    this.subjectiveColumns = new ArrayList<>(schedule.getSubjectiveStations());
     Collections.sort(scheduleData, TEAM_NUMBER_COMPARATOR);
   }
 
   /**
    * @see javax.swing.table.TableModel#getColumnCount()
    */
+  @Override
   public int getColumnCount() {
     return (JUDGE_COLUMN
         + 1)
-        + subjectiveColumns.size() + (schedule.getNumberOfRounds()
+        + subjectiveColumns.size()
+        + (schedule.getTotalNumberOfRounds()
             * NUM_COLUMNS_PER_ROUND);
   }
 
@@ -74,19 +76,16 @@ import javax.swing.table.AbstractTableModel;
 
   public int getFirstPerformanceColumn() {
     return JUDGE_COLUMN
-        + 1 + subjectiveColumns.size();
+        + 1
+        + subjectiveColumns.size();
   }
 
-  /**
-   * @see javax.swing.table.TableModel#getRowCount()
-   */
+  @Override
   public int getRowCount() {
     return scheduleData.size();
   }
 
-  /**
-   * @see javax.swing.table.TableModel#getValueAt(int, int)
-   */
+  @Override
   public Object getValueAt(final int rowIndex,
                            final int columnIndex) {
     final TeamScheduleInfo schedInfo = scheduleData.get(rowIndex);
@@ -107,17 +106,33 @@ import javax.swing.table.AbstractTableModel;
     } else {
       final int perfColIdx = columnIndex
           - getFirstPerformanceColumn();
-      final int round = perfColIdx
+
+      final int globalRoundIndex = perfColIdx
           / NUM_COLUMNS_PER_ROUND;
+
+      final PerformanceTime performance;
+      final int numPracticeRounds = schedule.getNumberOfPracticeRounds();
+      if (globalRoundIndex >= numPracticeRounds) {
+        // regular match play rounds are in columns after the practice rounds
+        final int roundIndex = globalRoundIndex
+            - numPracticeRounds;
+        performance = schedInfo.enumerateRegularMatchPlayPerformances().filter(p -> p.getRight() == roundIndex)
+                               .map(Pair::getLeft).findFirst().orElse(null);
+      } else {
+        // practice round
+        performance = schedInfo.enumeratePracticePerformances().filter(p -> p.getRight() == globalRoundIndex)
+                               .map(Pair::getLeft).findFirst().orElse(null);
+      }
+
       switch (perfColIdx
           % NUM_COLUMNS_PER_ROUND) {
       case 0:
       case 3:
-        return schedInfo.getPerfTime(round);
+        return performance.getTime();
       case 1:
-        return schedInfo.getPerfTableColor(round);
+        return performance.getTable();
       case 2:
-        return schedInfo.getPerfTableSide(round);
+        return performance.getSide();
       default:
         return null;
       }
@@ -133,14 +148,14 @@ import javax.swing.table.AbstractTableModel;
     } else if (columnIndex == JUDGE_COLUMN) {
       return String.class;
     } else if (columnIndex < getFirstPerformanceColumn()) {
-      return Date.class;
+      return LocalTime.class;
     } else {
       final int perfColIdx = columnIndex
           - getFirstPerformanceColumn();
       switch (perfColIdx
           % NUM_COLUMNS_PER_ROUND) {
       case 0:
-        return Date.class;
+        return LocalTime.class;
       case 1:
         return String.class;
       case 2:
@@ -161,18 +176,32 @@ import javax.swing.table.AbstractTableModel;
       return "Judge";
     } else if (columnIndex < getFirstPerformanceColumn()) {
       return subjectiveColumns.get(columnIndex
-          - JUDGE_COLUMN - 1);
+          - JUDGE_COLUMN
+          - 1);
     } else {
       final int perfColIdx = columnIndex
           - getFirstPerformanceColumn();
-      final int round = perfColIdx
-          / NUM_COLUMNS_PER_ROUND;
+
       switch (perfColIdx
           % NUM_COLUMNS_PER_ROUND) {
       case 0:
-        return "Perf #"
-            + (round
-                + 1);
+        final int globalRoundIndex = perfColIdx
+            / NUM_COLUMNS_PER_ROUND;
+
+        final int numPracticeRounds = schedule.getNumberOfPracticeRounds();
+        if (globalRoundIndex >= numPracticeRounds) {
+          // regular match play
+          final int roundIndex = globalRoundIndex
+              - numPracticeRounds;
+          return "Perf #"
+              + (roundIndex
+                  + 1);
+        } else {
+          // practice round
+          return "Practice #"
+              + (globalRoundIndex
+                  + 1);
+        }
       case 1:
         return "Table";
       case 2:
@@ -185,7 +214,7 @@ import javax.swing.table.AbstractTableModel;
 
   /**
    * Find the index of the specified team in this model.
-   * 
+   *
    * @param teamNumber
    * @return -1 if none found
    */