+++ Changes.md
@@ -1,3 +1,4 @@
+* #727 - Add support for practice rounds
 * #719 - Add judges room and performance area to the footer
 * #725 - rework subjective sheets to have more comment space
 * #726 - improve readability of subjective score summary